pyproject.toml
[project]
name = "dbt-data-pipeline-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.10.11",
    "dbt-databricks>=1.10.9",
    "dbt-python>=1.10.11",
]
